requirements.txt
opencv-python
numpy
Pillow
requests
pymilvus>=2.3.0